+++ CHANGELOG.md
@@ -2,12 +2,28 @@
 
 All changes to this project will be documented in this file.
 
-## [1.1] - 2026-02-28
+## [1.2.0] - 2026-03-01
+
+### Added
+
+- Added stat panel dynamically changing size based on the font size
+- Added font size options for enchants on the character panel and the inspect panel
+- Added font size options for the stats in the stats panel
+
+### Fixed
+
+- Fixed overlap of stats values
+
+### Changed
+
+- Changed pfUI detection
+
+## [1.1.0] - 2026-02-28
 
 ### Removed
 
 - Removed everything related to gear score
 
-## [1.0] - 2026-02-27
+## [1.0.0] - 2026-02-27
 
 Initial release